+++ addon/components/file-browser/component.js
@@ -136,6 +136,13 @@ export default Ember.Component.extend(Analytics, {
     browserState: Ember.computed('loaded', '_items', function() {
         return this.get('loaded') ? (this.get('_items').length ? (this.get('items').length ? 'show' : 'filtered') : 'empty') : 'loading';
     }),
+    clickable: Ember.computed('browserState', function() {
+        let clickable = [".dz-message"];
+        if (this.get('edit') && this.get('display').contains('upload-button') && this.get('browserState') == 'empty') {
+            clickable.push(".file-browser-list");
+        }
+        return clickable;
+    }),
     actions: {
         //dropzone listeners
         addedFile(_, __, file) {