+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Margins for scrollbar on `file-browser`
+- Clickability on dropzone widget
 
 ## [0.12.4] - 2017-12-04
 ### Added